src/requirements.txt
google-cloud-compute==1.15.0
google-cloud-securitycenter==1.25.0
google-cloud-logging==3.8.0
google-cloud-resource-manager==1.11.0
google-cloud-iam==2.16.0
google-cloud-pubsub==2.29.0
google-cloud-storage==2.19.0
google-cloud-monitoring==2.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-gcp-trace==1.6.0
functions-framework==3.4.0
